Add clap and tracing
