+++ Cargo.toml
@@ -4,9 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0.71"
 arrow = "42.0.0"
 bitcoin = "0.30.0"
 bitcoincore-rest = "2.0.0"
 chrono = "0.4.26"
+clap = { version = "4.3.10", features = ["derive"] }
 parquet = "42.0.0"
 tokio = { version = "1.28.2", features = ["macros", "full"] }
+tracing = "0.1.37"
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
